Add zoom dropdown
Adds a little dropdown next to the zoom slider.
The dropdown display the length current zoom "snippet".
This is supposed to give users a better sense of
how long the timeline actually is when zooming.
The dropdown also lets users skip to commonly used
zoom values (the "commonly used values" are arbitrarly defined by me).