Don't offer languages with little or no translation
The language selector offers quite the range of languages.
However, for most of them there are not actual translations,
so an eager user might be quite dissapointed when they
try to change to their favourite language and nothing happens.

This patch reduces the number of languages so that only
languages that currently have a translation percentage of 70% or
higher are still selectable.

Is there a better way to go about this than hardcoding which languages
are available? Probably. But this works for now.